feature: Use uuid.UUID for ExternalBlob in Record (#154)
The ExternalBlob field of the Record object has been changed to the uuid.UUID type because it's an internal-only key to identify blob reference objects. uuid.UUID is safer than string and guaranteed to be valid. This does not break compatibilities because it's not used yet.
This is part of #69 and #153.